# openapi.json
{
  "components": {
    "schemas": {
...
      "DraftRequest": {
        "properties": {
          "concept": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Concept",
            "type": "string"
          }
        },
        "required": [
          "concept"
        ],
        "title": "DraftRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NPC": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "hidden_agenda": {
            "default": "",
            "title": "Hidden Agenda",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "NPC",
        "type": "object"
      },
...
      "Role": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "private_info": {
            "default": "",
            "title": "Private Info",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "Role",
        "type": "object"
      },
...
      "ScenarioDraft": {
        "description": "LLM output schema for the AI-expand builder.",
        "properties": {
          "gm_notes": {
            "default": "",
            "title": "Gm Notes",
            "type": "string"
          },
          "goal": {
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "npcs": {
            "items": {
              "$ref": "#/components/schemas/NPC"
            },
            "maxItems": 10,
            "title": "Npcs",
            "type": "array"
          },
          "premise": {
            "minLength": 1,
            "title": "Premise",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Roles",
            "type": "array"
          },
          "setting": {
            "default": "",
            "title": "Setting",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "tone": {
            "default": "",
            "title": "Tone",
            "type": "string"
          }
        },
        "required": [
          "title",
          "premise",
          "goal",
          "roles"
        ],
        "title": "ScenarioDraft",
        "type": "object"
      },
      "ScenarioIn": {
        "properties": {
          "category": {
            "default": "",
            "maxLength": 100,
            "title": "Category",
            "type": "string"
          },
          "context_disclaimer": {
            "default": "",
            "maxLength": 4000,
            "title": "Context Disclaimer",
            "type": "string"
          },
          "context_enabled": {
            "default": false,
            "title": "Context Enabled",
            "type": "boolean"
          },
          "context_prompt": {
            "default": "",
            "maxLength": 4000,
            "title": "Context Prompt",
            "type": "string"
          },
          "gm_notes": {
            "default": "",
            "title": "Gm Notes",
            "type": "string"
          },
          "goal": {
            "default": "",
            "title": "Goal",
            "type": "string"
          },
          "npcs": {
            "items": {
              "$ref": "#/components/schemas/NPC"
            },
            "title": "Npcs",
            "type": "array"
          },
          "premise": {
            "default": "",
            "title": "Premise",
            "type": "string"
          },
          "risk_domain": {
            "default": "general",
            "enum": [
              "general",
              "health",
              "legal",
              "financial",
              "safety"
            ],
            "title": "Risk Domain",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "minItems": 1,
            "title": "Roles",
            "type": "array"
          },
          "setting": {
            "default": "",
            "title": "Setting",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "tone": {
            "default": "",
            "maxLength": 200,
            "title": "Tone",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "ScenarioIn",
        "type": "object"
      },
...
      "ScenarioOut": {
        "properties": {
          "category": {
            "default": "",
            "maxLength": 100,
            "title": "Category",
            "type": "string"
          },
          "context_disclaimer": {
            "default": "",
            "maxLength": 4000,
            "title": "Context Disclaimer",
            "type": "string"
          },
          "context_enabled": {
            "default": false,
            "title": "Context Enabled",
            "type": "boolean"
          },
          "context_prompt": {
            "default": "",
            "maxLength": 4000,
            "title": "Context Prompt",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "featured_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Featured Rank"
          },
          "gm_notes": {
            "default": "",
            "title": "Gm Notes",
            "type": "string"
          },
          "goal": {
            "default": "",
            "title": "Goal",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_library": {
            "default": false,
            "title": "Is Library",
            "type": "boolean"
          },
          "is_living": {
            "default": false,
            "title": "Is Living",
            "type": "boolean"
          },
          "is_premium": {
            "default": false,
            "title": "Is Premium",
            "type": "boolean"
          },
          "npcs": {
            "items": {
              "$ref": "#/components/schemas/NPC"
            },
            "title": "Npcs",
            "type": "array"
          },
          "premise": {
            "default": "",
            "title": "Premise",
            "type": "string"
          },
          "risk_domain": {
            "default": "general",
            "enum": [
              "general",
              "health",
              "legal",
              "financial",
              "safety"
            ],
            "title": "Risk Domain",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "minItems": 1,
            "title": "Roles",
            "type": "array"
          },
          "setting": {
            "default": "",
            "title": "Setting",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "tone": {
            "default": "",
            "maxLength": 200,
            "title": "Tone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "title",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "ScenarioOut",
        "type": "object"
      },
...
      "ScenarioUpdateOut": {
        "description": "Published situation-log entry, visible to players.",
        "properties": {
          "changes": {
            "title": "Changes",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "headline": {
            "title": "Headline",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "headline",
          "summary",
          "changes",
          "sources",
          "created_at"
        ],
        "title": "ScenarioUpdateOut",
        "type": "object"
      },
...
      "Source": {
        "properties": {
          "lean": {
            "default": "",
            "title": "Lean",
            "type": "string"
          },
          "outlet": {
            "title": "Outlet",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "outlet"
        ],
        "title": "Source",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CYOA Scenario Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scenarios": {
      "get": {
        "operationId": "list_scenarios",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioOut"
                  },
                  "title": "Response List Scenarios",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "scenarios"
        ]
      },
      "post": {
        "operationId": "create_scenario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/scenarios/draft": {
      "post": {
        "operationId": "draft_scenario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
...
    "/api/scenarios/{scenario_id}": {
      "delete": {
        "operationId": "delete_scenario",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Scenario",
        "tags": [
          "scenarios"
        ]
      },
      "get": {
        "operationId": "get_scenario",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenario",
        "tags": [
          "scenarios"
        ]
      },
      "put": {
        "operationId": "update_scenario",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
...
    "/api/scenarios/{scenario_id}/updates": {
      "get": {
        "description": "The player-facing situation log: published updates only, newest first.",
        "operationId": "list_scenario_updates",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioUpdateOut"
                  },
                  "title": "Response List Scenario Updates",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scenario Updates",
        "tags": [
          "scenarios"
        ]
      }
    }
  }
}